# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Short Breaks code test
# during dev, should point to your local machine
host: localhost:10010
basePath: /
# 
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /user/{id}:
    x-swagger-router-controller: "user"
    get:
      description: "Gets the user with the specified id"
      operationId: "getUser"
      produces:
      - "application/json"
      parameters:
        - name: "id"
          in: path
          description: "id of user to return"
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid ID"
        404:
          description: "User not found"
    delete:
      description: "Deletes the user with the specified id"
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
        - name: "id"
          in: path
          description: "id of user to delete"
          required: true
          type: string
      responses:
        200:
          description: Success
        400:
          description: "Invalid ID"
    put:
      description: "User details will be updated for the given user id"
      operationId: "updateUser"
      parameters:
        - name: "id"
          in: path
          description: "id of user to update"
          required: true
          type: string
        - name: "user"
          in: body
          description: "User object containing email, forename and surname. All these fields are optional."
          required: false
          schema:
            $ref: "#/definitions/User"
      responses:
        200:
          description: "Success. Full user object returned in body."
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid parameter"
  /user:
    x-swagger-router-controller: user
    post:
      description: "Adds a new user"
      operationId: "addUser"
      parameters:
        - name: "user"
          in: body
          description: "User object, in which the email, forename and surname should be set. Forename and surname should each be at least 3 characters long."
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        200:
          description: "Success. Location returned in header; full user object returned in body."
          headers:
            Location:
              type: string
              format: url
              description: "Relative URL of the new user"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid parameter"
  /users:
    x-swagger-router-controller: user
    get:
      description: "Returns a list of users, optionally filtered by field values"
      operationId: "listUsers"
      parameters:
        - name: "filterEmail"
          in: query
          type: string
          description: "Only list the user with email exactly matching"
          required: false
        - name: "filterForename"
          in: query
          type: string
          description: "Only list users with forenames exactly matching"
          required: false
        - name: "filterSurname"
          in: query
          type: string
          description: "Only list users with surnames exactly matching"
          required: false
      responses:
        200:
          description: "Success. List of users returned in body."
          schema:
            $ref: "#/definitions/UserList"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "string"
        readOnly: true
      email:
        type: "string"
      forename:
        type: "string"
      surname:
        type: "string"
      created:
        type: "string"
        format: "date-time"
        readOnly: true
  UserList:
    type: array
    items:
      $ref: '#/definitions/User'